Simplified workaround for terraform-docs

+++ terraform_docs.sh
@@ -25,36 +25,37 @@ main() {
     esac
   done
 
-  local hack_terraform_docs=$(terraform version | head -1 | grep -c 0.12)
+  local hack_terraform_docs
+  hack_terraform_docs=$(terraform version | head -1 | grep -c 0.12)
 
-  which terraform-docs 2>&1 >/dev/null || ( echo "terraform-docs is required"; exit 1)
-  local terraform_docs_version=$(terraform-docs version | head -1 | grep -E -o "([0-9]{1,}\.)+[0-9]{1,}")
+  if [[ ! $(command -v terraform-docs) ]]; then
+    echo "ERROR: terraform-docs is required by terraform_docs pre-commit hook but is not installed or in the system's PATH."
+    exit 1
+  fi
 
-  if [[ "$(check_terraform_docs_version "$terraform_docs_version")" == "1" ]]; then
+  local is_old_terraform_docs
+  is_old_terraform_docs=$(terraform-docs version | grep -o "v0.[1-7]" | tail -1)
+
+  if [[ -z "$is_old_terraform_docs" ]]; then  # Using terraform-docs 0.8+ (preferred)
 
     terraform_docs "0" "$args" "$files"
 
-  elif [[ "$hack_terraform_docs" == "1" ]]; then
-    which awk 2>&1 >/dev/null || ( echo "awk is required for terraform-docs hack to work with Terraform 0.12"; exit 1)
+  elif [[ "$hack_terraform_docs" == "1" ]]; then # Using awk script because terraform-docs is older than 0.8 and terraform 0.12 is used
+
+    if [[ ! $(command -v awk) ]]; then
+      echo "ERROR: awk is required for terraform-docs hack to work with Terraform 0.12."
+      exit 1
+    fi
 
     tmp_file_awk=$(mktemp "${TMPDIR:-/tmp}/terraform-docs-XXXXXXXXXX")
     terraform_docs_awk "$tmp_file_awk"
     terraform_docs "$tmp_file_awk" "$args" "$files"
     rm -f "$tmp_file_awk"
-  else
+
+  else # Using terraform 0.11 and no awk script is needed for that
+
     terraform_docs "0" "$args" "$files"
-  fi
 
-}
-
-check_terraform_docs_version() {
-  readonly currentver="$1"
-
-  requiredver="0.8.0"
-  if [[ "$(printf '%s\n' "$requiredver" "$currentver" | sort -V | head -n1)" = "$requiredver" ]]; then
-    echo "1"
-  else
-    echo "0"
   fi
 }
 
@@ -94,7 +95,7 @@ terraform_docs() {
     fi
 
   if [[ "$terraform_docs_awk_file" == "0" ]]; then
-    terraform-docs md $args ./ > "$tmp_file"
+    terraform-docs md "$args" ./ > "$tmp_file"
   else
     # Can't append extension for mktemp, so renaming instead
     tmp_file_docs=$(mktemp "${TMPDIR:-/tmp}/terraform-docs-XXXXXXXXXX")
@@ -102,7 +103,7 @@ terraform_docs() {
     tmp_file_docs_tf="$tmp_file_docs.tf"
 
     awk -f "$terraform_docs_awk_file" ./*.tf > "$tmp_file_docs_tf"
-    terraform-docs md $args "$tmp_file_docs_tf" > "$tmp_file"
+    terraform-docs md "$args" "$tmp_file_docs_tf" > "$tmp_file"
     rm -f "$tmp_file_docs_tf"
   fi
 
@@ -121,7 +122,7 @@ terraform_docs() {
 terraform_docs_awk() {
   readonly output_file=$1
 
-  cat <<"EOF" > $output_file
+  cat <<"EOF" > "$output_file"
 # This script converts Terraform 0.12 variables/outputs to something suitable for `terraform-docs`
 # As of terraform-docs v0.6.0, HCL2 is not supported. This script is a *dirty hack* to get around it.
 # https://github.com/segmentio/terraform-docs/
